Remove unneeded progressStatus.NextSegment. Fix indentation. Change AlphaPeptDeepLibraryBuilder to use CPU when running in simulation mode NONVIDIAHARD.
When running AlphapeptdeepBuildLibraryTest, run in simulation mode NONVIDIAHARD to prevent the test node from trying to install and use NVIDIA gpu (which requires admin elevation). Attempting this to fix Amazon test node failure, when NVIDIA card is present but we don't want to install any NVIDIA software.

Allow test to deletePython or keep it for faster testing.

+++ pwiz_tools/Skyline/TestPerf/AlphapeptdeepBuildLibraryTest.cs
@@ -36,10 +36,15 @@ namespace TestPerf
     [TestClass]
     public class AlphapeptdeepBuildLibraryTest : AbstractFunctionalTestEx
     {
-        [TestMethod]
+        // setting _deletePython to false allows the test to reuse existing installation
+        private bool _deletePython = true;
+
+        [TestMethod] 
         public void TestAlphaPeptDeepBuildLibrary()
         {
-            AssertEx.IsTrue(PythonInstaller.DeleteToolsPythonDirectory());
+            if (_deletePython) 
+                AssertEx.IsTrue(PythonInstaller.DeleteToolsPythonDirectory());
+            
             TestFilesZip = "TestPerf/AlphapeptdeepBuildLibraryTest.zip";
             RunFunctionalTest();
         }
@@ -121,6 +126,8 @@ namespace TestPerf
                 }
                 else
                 {
+                    PythonInstaller.SimulatedInstallationState =
+                        PythonInstaller.eSimulatedInstallationState.NONVIDIAHARD;
                     RunUI(() => { buildLibraryDlg.OkWizardPage(); });
                     WaitForClosedForm<BuildLibraryDlg>();
                 }
@@ -307,8 +314,7 @@ namespace TestPerf
             else // clickNo == null
             {
                 // Say 'Cancel'
-                RunDlg<AlertDlg>(nvidiaDlg.ClickCancel, confirmDlg => { ConfirmPythonSuccess(confirmDlg);
-                });
+                RunDlg<AlertDlg>(nvidiaDlg.ClickCancel, confirmDlg => { ConfirmPythonSuccess(confirmDlg); });
 
             }
